Fix condition for avoiding translating while upgrading
Previously I had tested for a nonexistent condition: If
$running_upgrade_script was set and false, do translation. But
$running_upgrade_script was only ever set to true.

The correct logic is to check whether the variable is set, and only
check its value if it is (expressed through a short-circuit boolean
operation).

Other places that do similar checks check for the existence of
globals, DOWN_FOR_MAINTENANCE or upgrade.php.