Maintain limitations on global trade log
Previously, the global trade log was limited to the same extent as other
members' trade logs. With the (optional) opening of trade logs, the
global log inadvertently was also included. Add a special check for this
log.

(There is a use case in that the member might want to filter their
personal log by date, but this would be better handled from that page
instead.)

Also, improve wording of the statement of limitations on the trade log.